changed naming

+++ newbank/src/main/java/se2/groupb/server/repository/AccountRepository.java
@@ -3,5 +3,5 @@ package se2.groupb.server.repository;
 import se2.groupb.server.account.Account;
 
 public interface AccountRepository {
-    Account findByID(int accountID);
+    Account findByID(Long accountID);
 }
